ref(ui): Handle general-docs redirect in Platform docs
There are currently two paths into the post-project 'configuration'
documentation:

  1. After setting up a project through the onboarding flow, this renders
     the 'Configure' component, which renders the 'Platform' component.
     Before this change the Configure component would handle redirecting
     to the generic 'getting-started' flow if no platform was slected.

  2. After setting up a project normally (through the create project
     buttons) we redirect directly to a route rendering the Platform
     component. Before this change, passing a 'other' platform would
     error out.

The redirect logic has moved from the Configure component to the
Platform component, to cover both these scenarios.

Fixes: GROW2-48
